+++ src/Hooks/useTotalTokenBalance/useTotalTokenBalance.ts
@@ -61,7 +61,7 @@ export const useTotalTokenBalance = (
     const tokenTotalToHumanFormatted = useMemo(() => {
         return BigNutils(tokenTotalBalance)
             .toHuman(token.decimals)
-            .toTokenFormat_string(decimals ?? 8, formatLocale)
+            .toTokenFormatFull_string(decimals ?? 8, formatLocale)
     }, [decimals, formatLocale, token.decimals, tokenTotalBalance])
 
     return { tokenTotalBalance, tokenTotalToHuman, tokenTotalToHumanFormatted, getGasFees }