+++ src/Screens/Flows/App/ActivityDetailsScreen/util/ActivityDetailsUtils.ts
@@ -45,6 +45,8 @@ export const getActivityTitle = (activity: Activity, LL: TranslationFunctions) =
             return LL.CONNECTED_APP_TITLE()
         case ActivityType.SIGN_TYPED_DATA:
             return LL.CONNECTED_APP_SIGN_TYPED_DATA()
+        case ActivityType.UNKNOWN_TX:
+            return LL.UNKNOWN_TX()
         default:
             warn(ERROR_EVENTS.ACTIVITIES, "Unknown activity type")
     }
@@ -69,6 +71,7 @@ export const getActivityModalTitle = (activity: Activity, LL: TranslationFunctio
         case ActivityType.TRANSFER_NFT:
         case ActivityType.CONNECTED_APP_TRANSACTION:
         case ActivityType.SIGN_TYPED_DATA:
+        case ActivityType.UNKNOWN_TX:
             return getActivityTitle(activity, LL)
         case ActivityType.B3TR_ACTION:
             return LL.B3TR_ACTION_MODAL_TITLE()